Remove validation from public API, change relevant test, add more docstrings, move iterator to its own file
